refactor: refactor formRegistration - move function to services
