+++ src/types/user/customer.ts
@@ -1,0 +1,10 @@
+export type Customer = {
+  version: number;
+  addresses: {
+    id: string;
+    streetName: string;
+    postalCode: string;
+    city: string;
+    country: string;
+  }[];
+};
